requirements.txt
Django==1.10.6
django-cors-headers==2.0.2
djangorestframework==3.5.4
djangorestframework-jwt==1.10.0
djoser==0.6.0
PyJWT==1.5.0
